only delete specified command messages

+++ events/guild/message.js
@@ -15,8 +15,13 @@ module.exports = async (bot, message) => {
   const commandfile = bot.commands.get(cmd) || bot.commands.get(bot.aliases.get(cmd));
   if (commandfile) {
     commandfile.run(bot, message, args);
+    console.log(commandfile.config.name);
     if (message.member.guild.me.hasPermission('MANAGE_MESSAGES')) {
-      message.delete();
+      const cmdToDel = ['help', 'invite', 'ping', 'support', 'uptime', 'say'];
+      console.log(cmdToDel.includes(commandfile.config.name));
+      if (cmdToDel.includes(commandfile.config.name)) {
+        message.delete();
+      }
     }
   }
